+++ lib/mongoid/persistable.rb
@@ -140,7 +140,7 @@ module Mongoid
       # @todo: Check if the document is persisted here.
       operations = yield({})
       persist_or_delay_atomic_operation(operations)
-      Threaded.clear_options!
+      clear_timeless_option
       self
     end