+++ lib/mongoid/threaded.rb
@@ -141,16 +141,6 @@ module Mongoid
       true
     end
 
-    # Clear out all options set on a one-time basis.
-    #
-    # @example Clear out the options.
-    #   Threaded.clear_options!
-    #
-    # @since 2.3.0
-    def clear_options!
-      self.timeless = nil
-    end
-
     # Exit autosaving a document on the current thread.
     #
     # @example Exit autosave.
@@ -353,44 +343,6 @@ module Mongoid
     # @since 2.1.0
     def scope_stack
       Thread.current["[mongoid]:scope-stack"] ||= {}
-    end
-
-    # Get the value of the one-off timeless call.
-    #
-    # @example Get the timeless value.
-    #   Threaded.timeless
-    #
-    # @return [ true, false ] The timeless setting.
-    #
-    # @since 2.3.0
-    def timeless
-      Thread.current["[mongoid]:timeless"]
-    end
-
-    # Set the value of the one-off timeless call.
-    #
-    # @example Set the timeless value.
-    #   Threaded.timeless = Band
-    #
-    # @param [ Class ] klass The document class.
-    #
-    # @since 2.3.0
-    def timeless=(klass)
-      Thread.current["[mongoid]:timeless"] = klass
-    end
-
-    # Is the current thread setting timestamps?
-    #
-    # @example Is the current thread timestamping?
-    #   Threaded.timestamping?(Band)
-    #
-    # @param [ Class ] klass The document class.
-    #
-    # @return [ true, false ] If timestamps can be applied.
-    #
-    # @since 2.3.0
-    def timestamping?(klass)
-      timeless != klass
     end
 
     # Is the document autosaved on the current thread?