+++ spec/mongoid/timestamps/timeless_spec.rb
@@ -4,6 +4,62 @@ describe Mongoid::Timestamps::Timeless do
 
   describe "#timeless" do
 
+    before(:all) do
+      class Chicken
+        include Mongoid::Document
+        include Mongoid::Timestamps
+
+        before_save :lay_timeless_egg
+
+        def lay_timeless_egg
+          Egg.timeless.create!
+        end
+      end
+
+      class Egg
+        include Mongoid::Document
+        include Mongoid::Timestamps
+      end
+    end
+
+    after(:all) do
+      Object.send(:remove_const, :Chicken)
+      Object.send(:remove_const, :Egg)
+    end
+
+    context "when others persist in the scope of the chain" do
+
+      context "when the root executes normally" do
+
+        let!(:chicken) do
+          Chicken.create!
+        end
+
+        it "creates the parent with a timestamp" do
+          expect(chicken.created_at).to_not be_nil
+        end
+
+        it "creates the child with no timestamp" do
+          expect(Egg.last.created_at).to be_nil
+        end
+      end
+
+      context "when the root executes timeless" do
+
+        let!(:chicken) do
+          Chicken.timeless.create!
+        end
+
+        it "creates the parent with a timestamp" do
+          expect(chicken.created_at).to be_nil
+        end
+
+        it "creates the child with no timestamp" do
+          expect(Egg.last.created_at).to be_nil
+        end
+      end
+    end
+
     context "when used as a proxy method" do
 
       context "when used on the document instance" do
@@ -25,7 +81,7 @@ describe Mongoid::Timestamps::Timeless do
         end
 
         it "clears out the timeless option after save" do
-          expect(Mongoid::Threaded.timeless).to be_false
+          expect(document).to be_timestamping
         end
 
         context "when subsequently persisting" do
@@ -55,7 +111,7 @@ describe Mongoid::Timestamps::Timeless do
         end
 
         it "clears out the timeless option after save" do
-          expect(Mongoid::Threaded.timeless).to be_false
+          expect(document).to be_timestamping
         end
 
         context "when subsequently persisting" do